Switch setup.sh from generic shell to bash. (#84)
Although this makes the script less portable, almost all users should have
access to bash on their machine. This also allows us to write more
readable .sh code, which makes maintaining this script eaiser.

I checked this script with `shellcheck` and ran it on a test project to
verify that the new changes work.

This should fix #83.