slack: Add color context to the body
After #134 [1] the slack message stopped having the vertical bar on the left,
which gave the message some context depending on the build status.

Passing an attachment block with color, has no effect to the sibling blocks.
To give blocks the proper color context, we need to move them inside the
attachments [2].

This must be done with having in mind that, according to Slack's
documentation [3] secondary attachments are deprecated.

However, there is no other way of giving color context to a block message [4].

[1] https://github.com/GoogleCloudPlatform/cloud-build-notifiers/pull/134
[2] https://api.slack.com/reference/messaging/attachments
[3] https://api.slack.com/messaging/composing/layouts#attachments
[4] https://api.slack.com/messaging/attachments-to-blocks#direct_equivalents