slack: Slack does not respect the predefined colors
Slack provides the ability to color the vertical bar when
creating an attachment with blocks. According to the documentation,
it also provides predefined colors [1].

However, the predefined colors see to work only with the legacy [2]
fields [2] and not with blocks.

I have not found anything in the documentation related to this but
can be easily verified using the kit builder [3].

I have two examples, one with an attachment block [4] and one with
an attachment text [5] that verify this.

[1] https://api.slack.com/reference/messaging/attachments#fields
[2] https://api.slack.com/reference/messaging/attachments#legacy_fields
[3] https://app.slack.com/block-kit-builder
[4] https://t.ly/ljPu
[5] https://t.ly/ub-R